+++ packages/demo-server/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @brok/demo-server
 
+## 0.5.0
+
+### Minor Changes
+
+- Update graph and dev deployments
+
+### Patch Changes
+
+- Updated dependencies
+  - @brok/sdk@0.7.0
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @brok/sdk@0.6.0
+
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @brok/sdk@0.5.0
+
+## 0.4.0
+
+### Minor Changes
+
+- New deployment and added properly did resolver
+
+### Patch Changes
+
+- Updated dependencies
+  - @brok/sdk@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes