+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @brok/sdk
 
+## 0.7.0
+
+### Minor Changes
+
+- Update graph and dev deployments
+
+### Patch Changes
+
+- Updated dependencies
+  - @brok/captable@0.7.0
+
+## 0.6.0
+
+### Minor Changes
+
+- Update deployment
+
+### Patch Changes
+
+- Updated dependencies
+  - @brok/captable@0.6.0
+
+## 0.5.0
+
+### Minor Changes
+
+- Added debug
+
+## 0.4.0
+
+### Minor Changes
+
+- New deployment and added properly did resolver
+
+### Patch Changes
+
+- Updated dependencies
+  - @brok/captable@0.5.0
+
 ## 0.3.3
 
 ### Patch Changes